print: Connect up prettyPrint() to Console.interact() and log()
It makes sense for the prettyPrint() function to be written in JS, but it
does need to be called internally in Console.interact() which is written
in C++. Therefore, create a slot on the global object in which to store
the prettyPrint() function, which can be retrieved when needing to call
the function from internal code.

Console.interact() no longer needs to use gjs_value_debug_string(), so
that code and its dependencies can be removed.

Also pass the arguments of log() and logError() through prettyPrint().

This enables the work from #107.