closure: Clean up gjs_closure_invoke()
Remove the return_exception argument from gjs_closure_invoke(). With this
argument, it was confusing whether an exception would be pending or not
after the closure invocation, and it was confusing whether the JS return
value represented an actual return value or an exception.

Now, gjs_closure_invoke() always follows the JSAPI return convention.

At the call site in function.cpp, with return_exception=true, previously
the returned value was re-thrown as an exception, only to be cleared again
and logged. Now it is just logged.

At the call site in value.cpp, with return_exception=false, previously the
exception would be logged in gjs_closure_invoke(), now we log it in the
calling code in closure_marshal().