object: Discard disposed GObject's and do not create wrappers for them
As per commit d37d6604 and commit c0003eb we may destroy an object
wrapper earlier than we used to do, this has some memory benefits but
we also need to make sure that we cleanup the object qdata so that it
never points to the just free'd object wrapper.

By doing this, however, we may end up re-associating a GObject that is
disposed but not yet finalized to another valid wrapper. This was allowed
before but wasn't really correct either because in such case it will be
just useless and won't be able to monitor its disposal again.

So, now during disposal notify save a qdata (that won't be removed
until finalization) on gobject marking it as disposed, then use this
data to prevent an object to be re-associated again to a valid wrapper.

In fact we just create a new wrapper that is marked as "disposed", and
so preventing most of actions to be performed.
We don't warn though, because this may just happen at the moment the
invalid object is incorrectly used.
As per this, we need to unset the object qdata only if the object is not
finalized, as we may still need to do it when it's finalized, or we'll
still crash when a disposed GObject will point to a wrapper that has been
already destroyed.

Added tests simulating this.

Fixes: #395