object: Never try to add a toggle reference on a disposed object
This should never happen as it would never be released afterwards, so
protect the call in all the cases, triggering a warning (but without
throwing as it's not a pure JS error) and return a bool so that we can
handle this case better (in fact if this happens we should not consider
the object toggle-reffed).

As per this we can now also warn about access to disposed object also
when trying to set a new expand property to a disposed object (adding
tests ensuring that this won't affect pre-set properties).