object: Also unset the object qdata during wrapper destruction
This codepath is hardly hit, but at this point we're sure that the
private-gjs we set as qdata is not valid anymore so it's safe to delete
it.

We've to do it only if the object has not been finalized though, not to
crash and not to break the assumptions for "disposed" gobject wrappers.

(cherry-picked from commit c746a668)