object: Catch finalized objects that are still under gjs scope
When an object is finalized its qdata is finally removed, so if while we
track the object in the wrapper the qdata is removed (after a disposal)
then we can be sure that the object is finalized.

We can so track this state by using a GDestroyNotify function as it will
be called with the wrapper pointer and so we can easily mark the object
as finalized in such case.

While the object is in the disposed state, waiting for finalization we
have to temporary give the role of monitoring the qdata to the disposed
quark so that during this phase the object has still the private qdata
unset, but still monitored for finalization, until we stop caring about
it.
If instead the object is created as a wrapper for a disposed object,
the private qdata still serves us to monitor its finalization without
having to override the disposed one, that may still be used by another
wrapper that is still living waiting for being garbage collected.

If this happens we can finally write proper critical messages and avoid
accessing to the object again.

Added a test to cover this case