object: Switch back to normal references on disposal
When an object is disposed we already remove the toggle reference on it
and we avoid that nothing will unref it again, however this can lead to
troubles when something else owns the last reference at garbage
collecting point if we didn't release the object earlier because in
~ObjectInstance we still expect to have a reference.

To avoid this, let's just keep it simpler and switch back to normal
reference again at dispose notify.

The ordering here is important though, as referencing the object may
cause a toggle-up event to happen, but we need to do it before removing
the toggle reference (as that could be the only one we have), however
after that we're still forcing a toggle-down event to happen, so we'll
still be unrooted afterwards.

Fixes: #399