installed-tests: Add GjsTestTools utilities to call native code for evil ops
At times we may need to simulate bad behaviors of other libraries, so to
check this we need to do evil hacks with the objects like ref or unref
them without having JS envolved.

To allow this, add an utility class that we introspect and can use for
testing.

(cherry-picked from commit 79c2ac3a9)