object: Cancel queued toggles on dispose notify
If when disposing an object we've toggle requests coming from another
thread (as it happens with GSettings) we need to cancel the queue as
the object won't be usable anymore and will be marked for being garbage
collected. This was currently causing us to hit a g_error, that isn't
actually problematic in this case because the object isn't going to be
usable anyways.

Add more test tools to simulate such case, and test it.

(cherry-picked from commit eb28da76)