object: Safely handle disposal when happening from other thread
In case we dispose an object from another thread we still queue a toggle
down event to happen in an idle, so that it can be picked by main
thread at next iteration.

However, in such case we'll have no more the instance set on the object
private data. We still keep track of it through the disposed data
though, so we can use it to trigger a toggle down event.

It's an error if we get an invalid ("true") pointer here, so we log
about it.