object: Do not disassociate a JS object if we have queued toggle up
In case an object has toggle up events queued we need them to happen
in order to make the object to be rooted again, and there's no need to
disassociate it the object.

If instead the object has queued toggle downs, we can instead just
cancel them as the object is unrooted anyways, so there's no issue in
proceeding.

Add test to check whether this is happening, we can simulate such
scenario by queuing a further ref on an object from another thread and
wait for GC to pick the object.

Fixes: #294