fundamental: Use value or instance pointer when no get/set funcs are provided
Some fundamental types are providing value set/get functions however
this is not mandatory as the fundamental can be just retrieved from the
GValue pointer or we can build a GValue using the instance and gtype.

A case for this are Gdk4 Events that does not provide such functions,
making impossible to use events signals.

Add tests simulating this, depending on a new gobject-introspection type
that respects this constrain.

Fixes: #398