arg: Dynamically get the gtype from JS object when none is provided
In some cases we're handing interfaces with unsed GType, a case is
GTypeInstance. When this happen we should rely on fallback marshaller
and try to figure out the actual C container for this type getting the
gtype from the JS object.

This allows to use g_value_init_from_instance() correctly between the
others cases.