gerror: Handle any value when converting thrown value to GError
In JS you can throw not only Error objects, but any value. If throwing a
value from a signal handler or vfunc with a GError** parameter, then that
value should be converted to a GError whether it is an Error object or
not.

Rename gjs_gerror_make_from_error() to gjs_gerror_make_from_thrown_value()
and refactor it to get the pending exception (thrown value) from the
context, and convert that into a GError.

If the object does not have a 'name' and 'message' property then it's
converted into a GJS_JS_ERROR_ERROR with an appropriate debug message; if
it does, then it's converted into the appropriate type of GJS_JS_ERROR.