testGObjectDestructionAccess: Ensure that disposed wrappers are unbound
When an object is disposed we discard its wrapper, if a new wrapper for
such objects is created it will stay connected to the object till the
wrapper is alive or the object is finalized.

However, we need to make sure that also the "disposed" wrapper object is
always unbound from its wrapped or we'll crash as in issue #395.

To test this, we can just repeatedly get from a C function a disposed
object so that we're going it to rewrap it over and over.