object: Block access to object only if it is finalized
For some time now we're warning whether a disposed object is used and
this is sane, however the data it holds may be still valid and in
general all the C code we access to is meant to be used after
disposition. While such code has not to be called when the object is
finalized.

Since now we properly track objects finalization we can go back to allow
most of the object actions (all but the signals connections, these
involve having toggle notifications that are not a tool we can use
at this point) with the actual object pointer.

In case the object is finalized instead we behave exactly as before.

Adapted tests and added ones for finalization case