context: Don't prepend installed GjsPrivate search path if uninstalled
We should set GJS_USE_UNINSTALLED_FILES (as the tests do) when hacking on
an uninstalled copy. Additionally, if this environment variable is set,
then we should not add the installed typelib location to the search path.

Otherwise this can lead to the confusing situation, if you are modifying
GjsPrivate, that an out-of-date installed typelib file can take precedence
over the one in the build directory, even when running the tests.