context: Ensure we check atomically if we're destroying from other threads
At each toggle notification coming from another thread we check if
we are destroying, however this value is set from the main thread (when
the final gc starts), so we have to access to this atomically.

Use std::atomic, and make ThreadSanitizer happy.

(cherry-picked from commit e9a9c0f5)