toggle: Keep track of objects via ObjectInstance not GObject's
While the ToggleQueue serves us to track GObject's using their pointers
is too fragile for us, because they can be also finalized from other
threads and we may not be able to track when this happens, leaving them
in the queue to be handled by the main thread, causing us to access to
invalid memory.

At the contrary, we totally control the life time of ObjectInstance's
and we're sure that they always are created and destroyed in the main
thread, so let's use them even inside the ToggleQueue.

In this way:
 - We can keep using the same strategy to enqueue them
 - We don't have to rely on monitoring GObject's finalization to clear
   the queue (that again may not happen in main thread).
 - When an object wrapper is destroyed, we are sure that we've removed
   the gobject toggle reference, and so that we'll never get another
   toggle event queued again till another wrapper is created.

Related to: #404