object: Rely on next ToggleQueue iteration to handle already queued toggle events
When handling the toggle events from the main thread, we may have
received a notification for a toggle event from multiple threads
concurrently causing us to consider this an error.

For example:
  Object 0x555555643b30 Toggle up, from main thread 1, queued down 1, up 1
  Object 0x555555643b30 Toggle down, from main thread 1, queued down 1, up 1
  Object 0x555555643b30 Toggle up, from main thread 0, queued down 0, up 0
  Object 0x555555643b30 Toggle up, from main thread 1, queued down 0, up 0
  Object 0x555555643b30 Toggle down, from main thread 0, queued down 0, up 0
  Object 0x555555643b30 Toggle down, from main thread 1, queued down 1, up 0

Till the introduction of locked operations during toggle notifications
we may not be sure how to handle this, but now we can be quite confident
that we're working with the actual toggle queue state and so when an
already-toggled event happens we can just be sure that it will be
handled at next idle, and so we can stop considering it an error.

To avoid having a queue of events that have no result, change the
enqueue code to cancel a previously queued event that is the opposite of
the one we're adding. We need to be able to add multiple events of the
same type though (as could come from different threads), so adapt the
cancellation code to this.

As it's just something we can't control our level, but depends on other
threads fired by applications. What we can do is only react properly,
and now we do.

Fixes: #297

(cherry-picked from commit f52496c7)