arg-cache: Do not mark a function returning void* as skip-all
While it's true that a void function has no return value, this is not
true in the case it's a void pointer.

Not doing this, will cause a crash at ffi level because we're passing to
it a null return value pointer, while it's expecting to fill it with
the returned data.

Fixes: #406