meson.build: Avoid putting two copies of libsysprof-capture-4.a in libgjs
`libgjs` was getting two copies of `libsysprof-capture-4.a` from:
```
  link_whole: libgjs_internal,
  dependencies: base_build_dep,
```
With regular dependencies this wouldn't happen (it would remain a list
of unique dependencies), but the `link_whole` introduced in ed6370b9061
has now forced it in a way that Meson can't deduplicate it. So linking of
`libgjs` was failing as the linker found two separate copies (with
separate addresses) of each symbol from `libsysprof-capture-4.a`.

To resolve this we now ensure that `profiler_deps` is only requested by
`libgjs_internal`, which itself is `link_whole`'d into `libgjs`. So
`libgjs` still gets `profiler_deps` but now only one copy.

What's special about this case is that it's a static library included in
a static library, so if that happens again then the same approach will be
needed.

Closes: https://gitlab.gnome.org/GNOME/gjs/-/issues/414