GObject: Call only setter for JS-defined construct-only properties
Only construct-only properties defined in JS GObject subclasses need
to poke into the several property accessors. This is notably superfluous
if the construct-only property comes from a parent GObject in C-land
(since then the JS object should not define the JS properties for it).

Avoid doing this with construct-only properties not defined by JS
objects.