build: Define WIN32_LEAN_AND_MEAN on Windows builds
This will fix the build by avoiding collisions with the Windows headers
when we use 'interface' in mem-private.h, as that is typically used in
COM programming on Windows.

Clean up the sources a bit by removing the #define WIN32_LEAN_AND_MEAN
as they are no longer needed in the sources as a result.