Add Internship Getting Started documentation
Iain from the SpiderMonkey team was kind enough to let me steal from the
document that Mozilla was distributing to Outreachy interns last year.
I've adapted it to describe our workflow.

To go with it, I've added a section on how to fork and clone the repo, to
the Hacking document, which I realized was missing.