object: Enqueue toggle ups when the heap is collecting
From my research it seems like .get() triggers an ExposeGCThingToActiveJS
call which can't be called while the heap is collecting - we guard against
that case in expose_to_js. Fix this by just queuing toggle ups if the heap
is collecting.

I'm pretty sure we're hitting this case:
https://gitlab.gnome.org/GNOME/glib/-/blob/main/gobject/gobject.c#L4635
I tested it and it seems like we're not getting a toggle up on the object
that is actually being finalized, so my only thought is perhaps it is an
object that the object getting temporarily toggled up holds a reference
to?

If that is the case, I think it makes sense to guard toggles while the
heap is collecting.