build: Fix building with gobject-introspection as a subproject
When building with gobject-introspection as a subproject, we want to find
the sources for the gobject-introspection tests in the subproject's source
tree, not installed. Querying the gobject-introspection dependency for a
'gidatadir' variable doesn't work if it's a subproject, because the
subproject doesn't export a variable by that name.

Instead, use the variables that it does export to locate the Regress and
GIMarshallingTests sources. (It doesn't currently export any variable for
WarnLib, so skip that test in the subproject case. I have opened a merge
request to add this to gobject-introspection.)

This is based on what PyGObject does.