build: Provide fallback location for GLib Valgrind suppressions
If building GLib as a subproject, it won't have a pkgconfig file to tell
where its install prefix is. It's most likely being built as a subproject
because the host system has an outdated version, not because it has no
version. Also, the user is most likely not running Valgrind. Given that,
it's probably OK to just fall back to the suppressions file in /usr.