esm/gi: Use __version__ property to check for conflicts
Currently the gi.versions object is not only used for specifying
which version to load, but also to catch repeated require() calls
with different versions.

The latter prevents falling back to a different version after the
original require() call fails, i.e. a pattern like the following:

```
let Soup;
try {
    Soup = gi.require('Soup', '3.0');
} catch (e) {
    Soup = gi.require('Soup', '2.4');
}
```

Address this by using the newly added __version__ property for
the version check, which reflects the actually loaded version
instead of the argument of a previous require() call.

It is also more reliable, as it works regardless of whether the
version was specified with the successful require() call:

```
const Gtk4 = gi.require('Gtk');
const Gtk3 = gi.require('Gtk', '3.0');
``

now throws an exception, while before both Gtk4 and Gtk3 were
resolved to the Gtk namespace (version 4.0).