debugger: Add option to ignore caught exceptions
Right now execution stops whenever an exception is thrown,
regardless of whether the exception was handled or not.

This gets annoying fast when exceptions are expected, for
example when handling EEXISTS after attempted file operations
instead of doing (racy) existence checks.

Address this by ignoring handled exceptions by default, with
the option to restore the old behavior.

https://gitlab.gnome.org/GNOME/gjs/-/issues/431