jsapi-util: Cast pointer to the correct type when calling ref/free-func
In GjsAutoPointer, T* is the type of the pointer, but F* is the type of
the pointer that ref_func/free_func accept. We need to cast to F* when
calling these functions.

Previously this mostly worked because F = void, but it would not compile
if F != void and F != T.