maint: Fix IWYU includes
The inline headers previously would not compile standalone, so they were
not being analyzed. Also check all IWYU pragmas to see if they are still
necessary; some of them were not.

The pragmas in gi/utils-inl.h are a result of CPPLINT requiring <utility>
for std::swap() even though it is also defined in <algorithm>, and IWYU
not picking up any of these three headers.