closure: Warn about unhandled promise rejections right away
In the new SpiderMonkey, the weirdness has been fixed where unregistering
a promise rejection when it is handled, actually works and does not
trigger the unhandled rejection warning. This means that a try-await-catch
construction is no longer treated as an unhandled promise rejection, so we
can warn earlier without causing false positives.

Closes: #417