object: Add null check for debug assertion
Without this, we may get a sudden crash when debug is enabled, if we are
typechecking an ObjectInstance with a null GObject pointer. This may occur
for example if you forget to do super._init() in your class's _init()
method, or it may be valid in some cases.

If it's a bug in your JS code, continuing on will give a better diagnostic
message than failing the assertion here.