build: Skip running minimal SpiderMonkey program on cross builds
We still try to build the program and get as far as we can even on cross
builds, to eliminate problems with the way paths are set up, for example.

For a cross build, we print a warning that builders should check that GJS
works before shipping it, since we are not performing that check now.