overrides: Support class fields and references in GObject.registerClass
This commit allows classes passed to GObject.registerClass
to reference themselves and use class fields.

```
GObject.registerClass(class MyObject {
    specialField = 5;

    doSomething() {
        console.log(this.specialField);
    }

    static new_special() {
        // Class self-references now work
        return new MyObject({ ... });
    }

});
```

Instead of copying descriptors to a new, native class in
registerClass, we can instead "wrap" the existing class
constructor in the relevant GObject mechanics.

Because constructors chain and can return arbitrary
objects, we rely on the GObject.Object constructor to
return a native ObjectInstance which the child class
can extend.

The native ObjectPrototype is stored in a private symbol
on the JavaScript prototype and is used to resolve
internally.

Fixes #331