jsapi-util: Remove return value of gjs_log_exception_full()
The documentation of gjs_log_exception_full() wasn't correct in that it
claimed to return false if no exception was pending, but it never actually
checked that. In fact it is used in logError() to log an exception that is
not pending, so this documentation was just incorrect.

Add the documentation instead to gjs_log_exception() where it belongs.