object: Fix interface property descriptor flags
Before the previous commit, the new property descriptor got the default
flags because JS::PropertyDescriptor didn't have a copy constructor, so it
was always configurable and non-enumerable. This changes the property
descriptor to be enumerable, since we generally define introspected API
properties to be enumerable.

JSPROP_SETTER and JSPROP_GETTER are already set by setGetterObject() and
setSetterObject(), so we don't need to provide them explicitly here.