jsapi-util: Improve BigInt debug functions
Using JS::ToBigInt64 to debug BigInt means that we can only access the
first 64-bit digit of the BigInt. (We can convert the whole thing to a
string, but that might allocate and fail, which we don't want in a debug
function.) Note this in debug messages, because I got myself confused a
few times while working on this branch.

Also use the new JS::BigIntIsNegative() API in mozjs91 to improve the
output, casting the first digit into a signed int if the BigInt was
negative, and an unsigned int if it was positive.