arg: Fix marshalling GDK_NONE as GdkAtom argument in
Mistakenly, gjs_value_to_g_argument() was expecting a JS Object value for
a GdkAtom, meaning that it would do a null check and throw if the value
was null while the argument was non-nullable. This broke the case of
passing in GDK_NONE (as the JS string 'NONE' or JS null) because they
correspond to the GdkAtom value of zero, which failed the null check.

The previous behaviour was mistakenly encoded in a test, so rewrite the
test as it ought to work.