arg-cache: Store less information for return type info
It's not necessary to have a whole GITypeInfo when we get the return
type of an argument, since this contains a lot of information that we
don't normally care about at this point.

We only need to determine a pointer where FFI can store the return value
and unpack that pointer into a GIArgument. For this we can get by with
only the GITypeTag, and additionally the GIInfoType in case of
GI_TYPE_TAG_INTERFACE. We store these in an Arg::ReturnTag structure,
which the argument cache returns (using Nothing in the case of a
function that doesn't have a return value.)

This will allow storing GITypeInfo pointers in fewer cases in the
argument cache, shrinking its size, in a future commit.

Use the new gi_type_tag_extract_ffi_return_value() in g-i 1.71 to do the
unpacking with only the information from Arg::ReturnTag, as previously
the API required passing in a GITypeInfo.

(Renamed from Arg::TypeTag and rewritten to use Maybe by Philip)