value: Fix regression converting Infinity and NaN to C integers
This regression has been around somewhat longer, but JS::ToInt32,
JS::ToInt64, and friends all treat Infinity and NaN as 0 when converting
to APIs that expect an integer. GJS used to do this as well. Add tests so
it doesn't regress again.

Additionally add tests for float and double, to make sure that Infinity
and NaN are preserved when converting to C float and double.