jsapi-util: Append cause when calling gjs_throw with exception pending
Previously, when calling gjs_throw() with an exception already pending,
the new exception would be dropped, after logging it in the debug output
in case it contained extra information.

Now that we have Error.cause, we can append the new exception to the cause
chain of the currently pending exception. This potentially gives more
relevant information immediately to the programmer without having to dig
in the debug log.

In the case where the end of the cause chain is not an object, or there is
a reference cycle in the cause chain, we fall back to the same debug
logging that we did before.