GObject: Support checking if a primitive is instance of an Interface
When using instanceof against an interface we assume that the instance
we pass is an object, but in case that's not the case, we're failing
while calling GObject.type_is_a instead of just returning false.

So, only perform such test in case we're handling an Object.

Add tests checking this case.

Fixes: #464