arg: Fix returning false without throwing exception
This g_return_val_if_fail() was incorrect because if
value_to_ghashtable_key() returns false, an exception must be pending.

However, the condition never fails because object property keys can only
be int32, string, or symbol; and JS_Enumerate() iterates over all
enumerable non-symbol property keys. At this point it's only possible to
have an int32 or string key, so change the check to an assertion, and add
a marshalling test ensuring that symbol keys are ignored.