context: Only warn about unhandled promises when the job queue is flushed
A promise rejection may be reported as not handled even though it's
going to be handled by a later added promise. In such case we were
correctly be informed about this via
unregister_unhandled_promise_rejection(), but given that since commit
4446ee7f we promptly warn about unhandled promise rejections our
unhandled rejections stack trace list was cleared earlier than we'd got
an unregistration event.

To avoid this, let's just wait for the job queue to be fully processed,
so that we won't accidentally clear the unhandled rejections list, until
those have been fully processed.

By doing this we are both avoiding misbehaving on the said case, and we
don't regress by keeping the exit-warns handled by #417 still valid.

Added tests to cover both a normal rejection case and the one that was
previously leading to a crash (assertion error).

Fixes: #466