package: Reverse order of running-from-source checks
The "run from source" feature has recently broken, though I'm not sure
exactly how recently. My best theory is that this is because Meson may
have changed how it organizes its build directory in recent versions, and
the new layout unfortunately means that the check in _runningFromSource()
succeeds under Meson, where before it did not.

Switching the order of _runningFromSource() and _runningFromMesonSource()
fixes the problem, since _runningFromMesonSource() relies on
Meson-specific environment variables and will definitely not succeed in a
non-Meson environment.