arg: Refactor gjs_g_array_new_for_type() to take GITypeTag
Rename it garray_new_for_storage_type() to reflect that we are supposed to
pass the result of g_type_info_get_storage_type(), and to remove the gjs_
prefix since it is a static function.

This allows calling this function in a future refactor where we don't have
access to the GITypeInfo pointer.