tests: Test conditions for returning FD list from async DBus method
An "Async" promise-based DBus method will return an FD list if it receives
one that is not falsy. This didn't match the behaviour of the "Remote"
callback-based DBus methods, which would always return either an FD list
or null as the third parameter to the callback.

So, when porting from "Remote" to "Async" you may get situations where a
null FD list is now undefined.

Add tests that encode this expectation.